requirements.txt
fontTools>=4.0.0
Pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
potrace>=0.1.0
scipy>=1.5.0 